Use workaround from nbudin to get python code
